--- autogoal/contrib/alphazero/MCTS.py
@@ -1,130 +1,0 @@
-import logging
-import math
-import numpy as np
-
-from .config import CPUCT, EPSILON, MCTS_SIMS
-
-log = logging.getLogger(__name__)
-
-
-class MCTS():
-    """
-    This class handles the MCTS tree.
-    """
-
-    def __init__(self, game, nnet):
-        self.game = game
-        self.nnet = nnet
-        self.Qsa = {}  # stores Q values for s,a (as defined in the paper)
-        self.Nsa = {}  # stores #times edge s,a was visited
-        self.Ns = {}  # stores #times board s was visited
-        self.Ps = {}  # stores initial policy (returned by neural net)
-
-        self.Es = {}  # stores game.getGameEnded ended for board s
-        self.Vs = {}  # stores game.getValidMoves for board s
-
-    def getActionProb(self, canonicalBoard, temp=1):
-        """
-        This function performs MCTS_SIMS number of simulations of MCTS starting from
-        canonicalBoard.
-
-        Returns:
-            probs: a policy vector where the probability of the ith action is proportional to Nsa[(s,a)]**(1./temp)
-        """
-        for i in range(MCTS_SIMS):
-            self.search(canonicalBoard)
-
-        s = self.game.stringRepresentation(canonicalBoard)
-        counts = [self.Nsa[(s, a)] if (s, a) in self.Nsa else 0 for a in range(self.game.getActionSize())]
-
-        if temp == 0:
-            bestAs = np.array(np.argwhere(counts == np.max(counts))).flatten()
-            bestA = np.random.choice(bestAs)
-            probs = [0] * len(counts)
-            probs[bestA] = 1
-            return probs
-
-        counts = [x ** (1. / temp) for x in counts]
-        counts_sum = float(sum(counts))
-        probs = [x / counts_sum for x in counts]
-        return probs
-
-    def search(self, canonicalBoard):
-        """
-        This function performs one iteration of MCTS. It is recursively called till a leaf node is found. At each node an action with max upper confidence bound is chosen.
-
-        Once a leaf node is found, the neural network is used to estimate an initial policy P and value v that gets propagated up the search path. In case the leaf node is a terminal state, the outcome is propagated up the search path. The values of Ns, Nsa, Qsa are
-        updated.
-
-        NOTE: the return values are the negative of the value of the current state. This is done since v is in [-1,1] and if v is the value of a state for the current player, then its value is -v for the other player.
-
-        Returns:
-            v: the negative of the value of the current canonicalBoard
-        """
-        
-        s = self.game.stringRepresentation(canonicalBoard)
-        
-
-        if s not in self.Es:
-            self.Es[s] = self.game.getGameEnded(canonicalBoard, 1)
-        if self.Es[s] != 0:
-            # terminal node
-            return -self.Es[s]
-
-        if s not in self.Ps:
-            # leaf node
-            self.Ps[s], v = self.nnet.predict(canonicalBoard)
-            
-            valids = self.game.getValidMoves(canonicalBoard, 1)
-            
-            self.Ps[s] = self.Ps[s] * valids  # masking invalid moves
-
-            sum_Ps_s = np.sum(self.Ps[s])
-           
-            if sum_Ps_s > 0:
-                self.Ps[s] /= sum_Ps_s  # renormalize              
-            else:
-                # if all valid moves were masked make all valid moves equally probable
-   
-                #log.error("All valid moves were masked, doing a workaround.")
-                self.Ps[s] = self.Ps[s] + valids
-                self.Ps[s] /= np.sum(self.Ps[s])
-
-            self.Vs[s] = valids
-            self.Ns[s] = 0
-            return -v
-
-        valids = self.Vs[s]
-        
-        cur_best = -99999999999
-        best_act = -1
-        
-        # pick the action with the highest upper confidence bound
-        for a in range(self.game.getActionSize()):
-            if valids[a]:
-                if (s, a) in self.Qsa:
-                    u = self.Qsa[(s, a)] + CPUCT * self.Ps[s][a] * math.sqrt(self.Ns[s]) / (
-                            1 + self.Nsa[(s, a)])
-                else:                   
-                    u = CPUCT * (self.Ps[s])[a] * math.sqrt(self.Ns[s] + EPSILON)
-
-                if u > cur_best:
-                    cur_best = u
-                    best_act = a
-
-        a = best_act
-        next_s, next_player = self.game.getNextState(canonicalBoard, 1, a)
-        next_b = self.game.getCanonicalForm(next_s, next_player)
-
-        v = self.search(next_b)
-
-        if (s, a) in self.Qsa:
-            self.Qsa[(s, a)] = (self.Nsa[(s, a)] * self.Qsa[(s, a)] + v) / (self.Nsa[(s, a)] + 1)
-            self.Nsa[(s, a)] += 1
-
-        else:
-            self.Qsa[(s, a)] = v
-            self.Nsa[(s, a)] = 1
-
-        self.Ns[s] += 1
-        return -v
