--- autogoal/contrib/alphazero/TicTacToeGame.py
@@ -1,210 +1,0 @@
-from __future__ import print_function
-from .GameInterface import Game
-import numpy as np
-
-"""
-Game class implementation for the game of TicTacToe.
-"""
-class TicTacToeGame(Game):
-    def __init__(self, n=3):
-        self.n = n
-        self._name = 'TicTacToe'
-
-    @property
-    def name(self):
-        return self._name
-
-    def getInitBoard(self):
-        # return initial board (numpy board)
-        b = Board(self.n)
-        return np.array(b.pieces)
-
-    def getBoardSize(self):
-        # (a,b) tuple
-        return (self.n, self.n)
-
-    def getActionSize(self):
-        # return number of actions
-        return self.n*self.n + 1
-
-    def getNextState(self, board, player, action):
-        # return the next (board, player) after player's action
-        # action must be a valid move
-        if action == self.n*self.n:
-            return (board, -player)
-        b = Board(self.n)
-        b.pieces = np.copy(board)
-        move = (int(action/self.n), action%self.n)
-        b.execute_move(move, player)
-        return (b.pieces, -player)
-
-    def getValidMoves(self, board, player):
-        # return a fixed size binary vector indicating the validity of every posible move
-        valids = [0]*self.getActionSize()
-        b = Board(self.n)
-        b.pieces = np.copy(board)
-        legalMoves =  b.get_legal_moves(player)
-        if len(legalMoves)==0:
-            valids[-1]=1
-            return np.array(valids)
-        for x, y in legalMoves:
-            valids[self.n*x+y]=1
-        return np.array(valids)
-
-    def getGameEnded(self, board, player):
-        # return 0 if not ended, 1 if player won, -1 if player lost
-        # player = 1
-        b = Board(self.n)
-        b.pieces = np.copy(board)
-
-        if b.is_win(player):
-            return 1
-        if b.is_win(-player):
-            return -1
-        if b.has_legal_moves():
-            return 0
-        # draw has a very little value 
-        return 1e-4
-
-    def getCanonicalForm(self, board, player):
-        # return board if player==1, else return -board if player==-1
-        return player*board
-
-    def getSymmetries(self, board, pi):
-        # mirror, rotational
-        assert(len(pi) == self.n**2+1)  # 1 for pass
-        pi_board = np.reshape(pi[:-1], (self.n, self.n))
-        l = []
-
-        for i in range(1, 5):
-            for j in [True, False]:
-                newB = np.rot90(board, i)
-                newPi = np.rot90(pi_board, i)
-                if j:
-                    newB = np.fliplr(newB)
-                    newPi = np.fliplr(newPi)
-                l += [(newB, list(newPi.ravel()) + [pi[-1]])]
-        return l
-
-    def stringRepresentation(self, board):
-        # 8x8 numpy array (canonical board)
-        return board.tostring()
-
-    @staticmethod
-    def display(board):
-        n = board.shape[0]
-
-        print("   ", end="")
-        for y in range(n):
-            print (y,"", end="")
-        print("")
-        print("  ", end="")
-        for _ in range(n):
-            print ("-", end="-")
-        print("--")
-        for y in range(n):
-            print(y, "|",end="")    # print the row #
-            for x in range(n):
-                piece = board[y][x]    # get the piece to print
-                if piece == -1: print("X ",end="")
-                elif piece == 1: print("O ",end="")
-                else:
-                    if x==n:
-                        print("-",end="")
-                    else:
-                        print("- ",end="")
-            print("|")
-
-        print("  ", end="")
-        for _ in range(n):
-            print ("-", end="-")
-        print("--")
-
-class Board():
-
-    # list of all 8 directions on the board
-    __directions = [(1,1),(1,0),(1,-1),(0,-1),(-1,-1),(-1,0),(-1,1),(0,1)]
-
-    def __init__(self, n=3):
-        "Set up initial board configuration."
-
-        self.n = n
-        # Create the empty board array.
-        self.pieces = [None]*self.n
-        for i in range(self.n):
-            self.pieces[i] = [0]*self.n
-
-        
-    # add [] indexer syntax to the Board
-    def __getitem__(self, index): 
-        return self.pieces[index]
-
-    def get_legal_moves(self, color):
-        """Returns all the legal moves for the given color.
-        (1 for white, -1 for black)
-        @param color not used and came from previous version.        
-        """
-        moves = set()  # store the legal moves.
-
-        # Get all the empty squares (color==0)
-        for y in range(self.n):
-            for x in range(self.n):
-                if self[x][y]==0:
-                    newmove = (x,y)
-                    moves.add(newmove)
-        return list(moves)
-
-    def has_legal_moves(self):
-        for y in range(self.n):
-            for x in range(self.n):
-                if self[x][y]==0:
-                    return True
-        return False
-    
-    def is_win(self, color):
-        """Check whether the given player has collected a triplet in any direction; 
-        @param color (1=white,-1=black)
-        """
-        win = self.n
-        # check y-strips
-        for y in range(self.n):
-            count = 0
-            for x in range(self.n):
-                if self[x][y]==color:
-                    count += 1
-            if count==win:
-                return True
-        # check x-strips
-        for x in range(self.n):
-            count = 0
-            for y in range(self.n):
-                if self[x][y]==color:
-                    count += 1
-            if count==win:
-                return True
-        # check two diagonal strips
-        count = 0
-        for d in range(self.n):
-            if self[d][d]==color:
-                count += 1
-        if count==win:
-            return True
-        count = 0
-        for d in range(self.n):
-            if self[d][self.n-d-1]==color:
-                count += 1
-        if count==win:
-            return True
-        
-        return False
-
-    def execute_move(self, move, color):
-        """Perform the given move on the board; 
-        color gives the color of the piece to play (1=white,-1=black)
-        """
-
-        (x,y) = move
-
-        # Add the piece to the empty square.
-        assert self[x][y] == 0
-        self[x][y] = color
