+++ autogoal/experimental/alphazero/_semantics.py
